remove root.go
