+++ cmd/config.go
@@ -28,7 +28,7 @@ import (
 )
 
 // configCmd represents the config command
-var configCmd = &cobra.Command{
+var ConfigCmd = &cobra.Command{
 	Use:   "config",
 	Short: "A brief description of your command",
 	Long: `A longer description that spans multiple lines and likely contains examples
@@ -43,7 +43,7 @@ to quickly create a Cobra application.`,
 }
 
 func init() {
-	rootCmd.AddCommand(configCmd)
+	// rootCmd.AddCommand(configCmd)
 
 	// Here you will define your flags and configuration settings.
