+++ cmd/create.go
@@ -28,7 +28,7 @@ import (
 )
 
 // createCmd represents the create command
-var createCmd = &cobra.Command{
+var CreateCmd = &cobra.Command{
 	Use:   "create",
 	Short: "A brief description of your command",
 	Long: `A longer description that spans multiple lines and likely contains examples
@@ -43,7 +43,7 @@ to quickly create a Cobra application.`,
 }
 
 func init() {
-	configCmd.AddCommand(createCmd)
+	ConfigCmd.AddCommand(CreateCmd)
 
 	// Here you will define your flags and configuration settings.
