+++ cmd/version.go
@@ -43,7 +43,7 @@ to quickly create a Cobra application.`,
 }
 
 func init() {
-	rootCmd.AddCommand(VersionCmd)
+	// .AddCommand(VersionCmd)
 
 	// Here you will define your flags and configuration settings.
